Add squish
Squish is a tool to pack many individual Lua scripts and their
modules into a single Lua script.

Squish has its homepage at

	http://matthewwild.co.uk/projects/squish/home

Squish is MIT-licensed. Note that the MIT license is GPL-compatible.

Currently we are generating wbgen2 in a non-portable way. Adding
squish is a first step towards doing this properly.

Signed-off-by: Emilio G. Cota <cota@braap.org>

git-svn-id: http://svn.ohwr.org/wishbone-gen@6 4537843c-45c2-4d80-8546-c3283569414f